refactor: give ssh-tmux lifecycle contracts dedicated files
