+++ Sources/RemoteTmuxMirrorTeardownReason.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+enum RemoteTmuxMirrorTeardownReason: Sendable, Equatable {
+    case sessionEnded
+    case explicitDetach
+}
